cacheable - chore: clean up types and enums
