+++ packages/cacheable/src/enums.ts
@@ -1,0 +1,17 @@
+export enum CacheableHooks {
+	BEFORE_SET = "BEFORE_SET",
+	AFTER_SET = "AFTER_SET",
+	BEFORE_SET_MANY = "BEFORE_SET_MANY",
+	AFTER_SET_MANY = "AFTER_SET_MANY",
+	BEFORE_GET = "BEFORE_GET",
+	AFTER_GET = "AFTER_GET",
+	BEFORE_GET_MANY = "BEFORE_GET_MANY",
+	AFTER_GET_MANY = "AFTER_GET_MANY",
+	BEFORE_SECONDARY_SETS_PRIMARY = "BEFORE_SECONDARY_SETS_PRIMARY",
+}
+
+export enum CacheableEvents {
+	ERROR = "error",
+	CACHE_HIT = "cache:hit",
+	CACHE_MISS = "cache:miss",
+}
